Functional sign-up form
It contains a sidebar with a logo and a title using a custom font.
At the bottom there are links to the photo author.

There is also a description to the right of the image. Below this
description, there is a form with some text elements: first name, last
name, email, phone number, twice for password and a submit button.

This input elements have some requisites like entering a valid name
for both first and last name, a standard US phone number and a password
with a number, an uppercase letter, a lowercase letter and at least 8
characters. This is done with the pattern attribute. There is also a
title attribute to describe what is expected from each input for the
user.

Underneath the form, there is a text with a link to simulate when the
user has an account and wants to log-in.

For the CSS part, almost everything has its attributes. The positioning
of elements is mainly done with flexbox and some parts with another
type of positioning. There are two fonts used for the text on the page.
Some pseudo-classes are used to give certain styles to the forms, like
:focus and :invalid, that applies when the user enters invalid data on
the inputs.